initial halo 3 mcc support
everything seems to work, poking had to be changed again to accomodate, which affects all mcc games going forward.

+++ src/Blamite/Blam/ThirdGen/ThirdGenCacheFile.cs
@@ -43,7 +43,7 @@ namespace Blamite.Blam.ThirdGen
 			_endianness = reader.Endianness;
 			_buildInfo = buildInfo;
 			_segmenter = new FileSegmenter(buildInfo.SegmentAlignment);
-			_expander = new ThirdGenPointerExpander((uint)buildInfo.ExpandMagic);
+			_expander = new ThirdGenPointerExpander(buildInfo.ExpandMagic);
 			Allocator = new MetaAllocator(this, 0x10000);
 			Load(reader, buildString);
 		}
